changed the reviewer phone number to activerecord datatype bigint so it can handle phone nums

+++ textbub_app/db/migrate/20160331002218_create_reviews.rb
@@ -4,7 +4,7 @@ class CreateReviews < ActiveRecord::Migration
       t.references :user
       t.string :content
       t.string :service_providers_name
-      t.integer :review_phone
+      t.bigint :review_phone
     end
   end
 end
